# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Skills4Attendance API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /allocation/allocate:
    post:
      operationId: allocateLearners
      tags: [allocation]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AllocationInput"
      responses:
        "200":
          description: Allocation applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AllocationResult"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /allocation/history:
    get:
      operationId: listAllocationHistory
      tags: [allocation]
      parameters:
        - name: learnerId
          in: query
          schema: { type: integer }
        - name: cohortId
          in: query
          schema: { type: integer }
        - name: tutorId
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Allocation history
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AllocationHistoryEntry" }

  # ---------- Attendance ----------
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    AllocationHistoryEntry:
      type: object
      required:
        [
          id,
          learnerId,
          learnerName,
          previousTutorId,
          previousTutorName,
          newTutorId,
          newTutorName,
          previousCohortId,
          previousCohortName,
          newCohortId,
          newCohortName,
          effectiveDate,
          transferReason,
          changedBy,
          changedByName,
          changedDate,
        ]
      properties:
        id: { type: integer }
        learnerId: { type: integer }
        learnerName: { type: string }
        previousTutorId: { type: ["integer", "null"] }
        previousTutorName: { type: ["string", "null"] }
        newTutorId: { type: ["integer", "null"] }
        newTutorName: { type: ["string", "null"] }
        previousCohortId: { type: ["integer", "null"] }
        previousCohortName: { type: ["string", "null"] }
        newCohortId: { type: ["integer", "null"] }
        newCohortName: { type: ["string", "null"] }
        effectiveDate: { type: string, format: date }
        transferReason: { type: ["string", "null"] }
        changedBy: { type: integer }
        changedByName: { type: string }
        changedDate: { type: string, format: date-time }
...
    AllocationInput:
      type: object
      required: [learnerIds, effectiveDate]
      properties:
        learnerIds:
          type: array
          items: { type: integer }
          minItems: 1
        tutorId: { type: integer }
        cohortId: { type: integer }
        effectiveDate: { type: string, format: date }
        transferReason: { type: string }
    AllocationResult:
      type: object
      required: [updated]
      properties:
        updated: { type: integer }
...